+++ android/Staccato_AN/app/src/main/java/com/on/staccato/domain/model/NicknameState.kt
@@ -1,0 +1,14 @@
+package com.on.staccato.domain.model
+
+sealed interface NicknameState {
+    data object Empty : NicknameState
+
+    data class Valid(val nickname: String) : NicknameState
+
+    data object InvalidFormat : NicknameState
+
+    data class InvalidLength(
+        val min: Int,
+        val max: Int,
+    ) : NicknameState
+}
